+++ src/constants.ts
@@ -100,6 +100,9 @@ export const CRANK_ACCOUNT_LIMIT = 12;
 export const CRANK_PERP_ACCOUNT_LIMIT = 10;
 export const MAX_MARKETS_TO_FETCH = 50;
 
+export const MIN_LOT_SIZE = 0.001;
+export const PERP_MARKET_ORDER_SPOT_SLIPPAGE = 0.02;
+
 // This is the most we can load per iteration without
 // hitting the rate limit.
 export const MARKET_LOAD_LIMIT = 12;