+++ src/risk.ts
@@ -9,7 +9,11 @@ import {
 import { assetToIndex, fromProgramAsset } from "./assets";
 import { Asset } from "./constants";
 import { assets, Decimal } from ".";
-import { calculateProductMargin, collectRiskMaps } from "./risk-utils";
+import {
+  calculateLiquidationPrice,
+  calculateProductMargin,
+  collectRiskMaps,
+} from "./risk-utils";
 
 export class RiskCalculator {
   private _marginRequirements: Map<Asset, Array<types.MarginRequirement>>;
@@ -63,7 +67,7 @@ export class RiskCalculator {
     );
   }
 
-  public calculateMarginRequirement(
+  private calculateMarginRequirement(
     marginRequirement: types.MarginRequirement,
     size: number,
     marginType: types.MarginType
@@ -109,15 +113,25 @@ export class RiskCalculator {
     return sideMultiplier * openingSize;
   }
 
+  public calculateUnpaidFunding(
+    account: MarginAccount,
+    accountType: types.ProgramAccountType
+  ): number;
+  public calculateUnpaidFunding(
+    account: CrossMarginAccount,
+    accountType: types.ProgramAccountType
+  ): Map<Asset, number>;
+
   /**
-   * Returns the unpaid funding for a given margin account.
-   * @param account the user's spread (returns 0) or margin account.
+   * Returns the unpaid funding for a given margin or crossmargin account.
+   * @param account The MarginAccount or CrossMarginAccount itself
+   * @param accountType MarginAccount or CrossMarginAccount
    */
   public calculateUnpaidFunding(
     account: any,
     accountType: types.ProgramAccountType = types.ProgramAccountType
       .MarginAccount
-  ): number | Map<Asset, number> {
+  ): any {
     // Spread accounts cannot hold perps and therefore have no unpaid funding
     if (accountType == types.ProgramAccountType.SpreadAccount) {
       return 0;
@@ -165,16 +179,80 @@ export class RiskCalculator {
   }
 
   /**
-   * Returns the unrealized pnl for a given margin or spread account.
-   * @param account the user's spread or margin account.
+   * Calculates the estimated realized PnL of the account by passing in an execution price and using taker fees
+   * @param account The MarginAccount or CrossMarginAccount itself
+   * @param accountType MarginAccount or CrossMarginAccount
+   * @param executionInfo Information about how the PnL is hypothetically realised - execution price, asset and whether to add fees
+   */
+  public estimateRealizedPnl(
+    account: any,
+    accountType: types.ProgramAccountType = types.ProgramAccountType
+      .MarginAccount,
+    executionInfo: { asset: Asset; price: number; addTakerFees: boolean }
+  ): number {
+    // Number for MarginAccount, Map for CrossMarginAccount
+    if (accountType == types.ProgramAccountType.CrossMarginAccount) {
+      let pnl = this.calculatePnl(
+        account as CrossMarginAccount,
+        accountType,
+        executionInfo
+      );
+      return Array.from(pnl.values()).reduce((a, b) => a + b, 0);
+    } else if (accountType == types.ProgramAccountType.MarginAccount) {
+      return this.calculatePnl(
+        account as MarginAccount,
+        accountType,
+        executionInfo
+      );
+    }
+
+    return 0;
+  }
+
+  public calculateUnrealizedPnl(
+    account: MarginAccount,
+    accountType: types.ProgramAccountType
+  ): number;
+  public calculateUnrealizedPnl(
+    account: CrossMarginAccount,
+    accountType: types.ProgramAccountType
+  ): Map<Asset, number>;
+
+  /**
+   * Calculates the unrealised PnL of the account, marking all positions to the markPrice
+   * @param account The MarginAccount or CrossMarginAccount itself
+   * @param accountType MarginAccount or CrossMarginAccount
    */
   public calculateUnrealizedPnl(
     account: any,
     accountType: types.ProgramAccountType = types.ProgramAccountType
+      .MarginAccount
+  ): any {
+    return this.calculatePnl(account, accountType, undefined);
+  }
+
+  private calculatePnl(
+    account: MarginAccount,
+    accountType: types.ProgramAccountType,
+    executionInfo:
+      | { asset: Asset; price: number; addTakerFees: boolean }
+      | undefined
+  ): number;
+  private calculatePnl(
+    account: CrossMarginAccount,
+    accountType: types.ProgramAccountType,
+    executionInfo:
+      | { asset: Asset; price: number; addTakerFees: boolean }
+      | undefined
+  ): Map<Asset, number>;
+  private calculatePnl(
+    account: any,
+    accountType: types.ProgramAccountType = types.ProgramAccountType
       .MarginAccount,
-    executionPrice: number = undefined,
-    addTakerFees: boolean = false
-  ): number | Map<Asset, number> {
+    executionInfo:
+      | { asset: Asset; price: number; addTakerFees: boolean }
+      | undefined = undefined
+  ): any {
     let pnl = 0;
 
     let i_list = [];
@@ -219,18 +297,24 @@ export class RiskCalculator {
       }
 
       let subExchange = Exchange.getSubExchange(asset);
+      let price =
+        executionInfo && executionInfo.asset == asset
+          ? executionInfo.price
+          : subExchange.getMarkPrice();
       if (size > 0) {
         assetPnl +=
-          convertNativeLotSizeToDecimal(size) *
-            (executionPrice ? executionPrice : subExchange.getMarkPrice()) -
+          convertNativeLotSizeToDecimal(size) * price -
           convertNativeBNToDecimal(position.costOfTrades);
       } else {
         assetPnl +=
-          convertNativeLotSizeToDecimal(size) *
-            (executionPrice ? executionPrice : subExchange.getMarkPrice()) +
+          convertNativeLotSizeToDecimal(size) * price +
           convertNativeBNToDecimal(position.costOfTrades);
       }
-      if (addTakerFees) {
+      if (
+        executionInfo &&
+        executionInfo.asset == asset &&
+        executionInfo.addTakerFees
+      ) {
         assetPnl -=
           convertNativeLotSizeToDecimal(Math.abs(size)) *
           (convertNativeBNToDecimal(Exchange.state.nativeD1TradeFeePercentage) /
@@ -248,18 +332,29 @@ export class RiskCalculator {
     }
   }
 
+  public calculateTotalInitialMargin(
+    account: MarginAccount,
+    accountType: types.ProgramAccountType,
+    skipConcession: boolean
+  ): number;
+  public calculateTotalInitialMargin(
+    account: CrossMarginAccount,
+    accountType: types.ProgramAccountType,
+    skipConcession: boolean
+  ): Map<Asset, number>;
   /**
    * Returns the total initial margin requirement for a given account.
    * This includes initial margin on positions which is used for
    * Place order, withdrawal and force cancels
-   * @param marginAccount   the user's MarginAccount.
+   * @param account the user's margin or crossmargin account.
+   * @param accountType MarginAccount or CrossMarginAccount
    */
   public calculateTotalInitialMargin(
     account: any,
     accountType: types.ProgramAccountType = types.ProgramAccountType
       .MarginAccount,
     skipConcession: boolean = false
-  ): number | Map<Asset, number> {
+  ): any {
     // Margin account only has 1 asset, but for CrossMarginAccount we iterate through all assets
     let assetList =
       accountType == types.ProgramAccountType.MarginAccount
@@ -333,18 +428,26 @@ export class RiskCalculator {
     }
   }
 
+  public calculateTotalMaintenanceMargin(
+    account: MarginAccount,
+    accountType: types.ProgramAccountType
+  ): number;
+  public calculateTotalMaintenanceMargin(
+    account: CrossMarginAccount,
+    accountType: types.ProgramAccountType
+  ): Map<Asset, number>;
   /**
    * Returns the total maintenance margin requirement for a given account.
    * This only uses maintenance margin on positions and is used for
    * liquidations.
-   * @param asset           underlying asset type
-   * @param marginAccount   the user's MarginAccount.
+   * @param account the user's margin or crossmargin account.
+   * @param accountType MarginAccount or CrossMarginAccount
    */
   public calculateTotalMaintenanceMargin(
     account: any,
     accountType: types.ProgramAccountType = types.ProgramAccountType
       .MarginAccount
-  ): number | Map<Asset, number> {
+  ): any {
     // Margin account only has 1 asset, but for CrossMarginAccount we iterate through all assets
     let assetList =
       accountType == types.ProgramAccountType.MarginAccount
@@ -383,18 +486,26 @@ export class RiskCalculator {
     }
   }
 
+  public calculateTotalMaintenanceMarginIncludingOrders(
+    account: MarginAccount,
+    accountType: types.ProgramAccountType
+  ): number;
+  public calculateTotalMaintenanceMarginIncludingOrders(
+    account: CrossMarginAccount,
+    accountType: types.ProgramAccountType
+  ): Map<Asset, number>;
   /**
    * Returns the total maintenance margin requirement for a given account including orders.
    * This calculates maintenance margin across all positions and orders.
    * This value is used to determine margin when sending a closing order only.
-   * @param asset           underlying asset type
-   * @param marginAccount   the user's MarginAccount.
+   * @param account the user's margin or crossmargin account.
+   * @param accountType MarginAccount or CrossMarginAccount
    */
   public calculateTotalMaintenanceMarginIncludingOrders(
     account: any,
     accountType: types.ProgramAccountType = types.ProgramAccountType
       .MarginAccount
-  ): number {
+  ): number | Map<Asset, number> {
     // Margin account only has 1 asset, but for CrossMarginAccount we iterate through all assets
     let assetList =
       accountType == types.ProgramAccountType.MarginAccount
@@ -402,6 +513,7 @@ export class RiskCalculator {
         : Exchange.assets;
 
     let margins = 0;
+    let marginMap: Map<Asset, number> = new Map();
     for (var asset of assetList) {
       const ledger =
         accountType == types.ProgramAccountType.MarginAccount
@@ -410,6 +522,8 @@ export class RiskCalculator {
       let size = ledger.position.size.toNumber();
       let bidOpenOrders = ledger.orderState.openingOrders[0].toNumber();
       let askOpenOrders = ledger.orderState.openingOrders[1].toNumber();
+
+      marginMap.set(asset, 0);
       if (bidOpenOrders == 0 && askOpenOrders == 0 && size == 0) {
         continue;
       }
@@ -423,7 +537,7 @@ export class RiskCalculator {
         shortLots += Math.abs(convertNativeLotSizeToDecimal(size));
       }
 
-      margins +=
+      let assetMargin =
         this.getPerpMarginRequirement(
           asset,
           // Positive for buys.
@@ -436,27 +550,30 @@ export class RiskCalculator {
           -shortLots,
           types.MarginType.MAINTENANCE
         );
+
+      margins += assetMargin;
+      marginMap.set(asset, assetMargin);
     }
 
-    return margins;
+    if (accountType == types.ProgramAccountType.MarginAccount) {
+      return margins;
+    } else {
+      return marginMap;
+    }
   }
 
   /**
    * Returns the aggregate margin account state.
-   * @param crossMarginAccount   the user's CrossMarginAccount.
+   * @param marginAccount   the user's CrossMarginAccount.
    */
   public getCrossMarginAccountState(
-    marginAccount: CrossMarginAccount,
-    pnlExecutionPrice: number = undefined,
-    pnlAddTakerFees: boolean = false
+    marginAccount: CrossMarginAccount
   ): types.CrossMarginAccountState {
     let balance = convertNativeBNToDecimal(marginAccount.balance);
     let accType = types.ProgramAccountType.CrossMarginAccount;
     let unrealizedPnl = this.calculateUnrealizedPnl(
       marginAccount,
-      accType,
-      pnlExecutionPrice,
-      pnlAddTakerFees
+      accType
     ) as Map<Asset, number>;
     let unpaidFunding = this.calculateUnpaidFunding(
       marginAccount,
@@ -464,7 +581,8 @@ export class RiskCalculator {
     ) as Map<Asset, number>;
     let initialMargin = this.calculateTotalInitialMargin(
       marginAccount,
-      accType
+      accType,
+      false
     ) as Map<Asset, number>;
     let initialMarginSkipConcession = this.calculateTotalInitialMargin(
       marginAccount,
@@ -475,6 +593,11 @@ export class RiskCalculator {
       marginAccount,
       accType
     ) as Map<Asset, number>;
+    let maintenanceMarginIncludingOrders =
+      this.calculateTotalMaintenanceMarginIncludingOrders(
+        marginAccount,
+        accType
+      ) as Map<Asset, number>;
 
     let upnlTotal = Array.from(unrealizedPnl.values()).reduce(
       (a, b) => a + b,
@@ -492,28 +615,36 @@ export class RiskCalculator {
       (a, b) => a + b,
       0
     );
+    let mmioTotal = Array.from(
+      maintenanceMarginIncludingOrders.values()
+    ).reduce((a, b) => a + b, 0);
 
     let availableBalanceInitial: number =
       balance + upnlTotal + unpaidFundingTotal - imTotal;
     let availableBalanceWithdrawable: number =
-      balance + upnlTotal + unpaidFundingTotal - imSkipConcessionTotal;
+      balance + upnlTotal + unpaidFundingTotal;
     let availableBalanceMaintenance: number =
       balance + upnlTotal + unpaidFundingTotal - mmTotal;
+    let availableBalanceMaintenanceIncludingOrders: number =
+      balance + upnlTotal + unpaidFundingTotal - mmioTotal;
     return {
       balance,
       availableBalanceInitial,
       availableBalanceMaintenance,
+      availableBalanceMaintenanceIncludingOrders,
       availableBalanceWithdrawable,
       assetState: collectRiskMaps(
         initialMargin,
         initialMarginSkipConcession,
         maintenanceMargin,
+        maintenanceMarginIncludingOrders,
         unrealizedPnl,
         unpaidFunding
       ),
       initialMarginTotal: imTotal,
       initalMarginSkipConcessionTotal: imSkipConcessionTotal,
       maintenanceMarginTotal: mmTotal,
+      maintenanceMarginIncludingOrdersTotal: mmioTotal,
       unrealizedPnlTotal: upnlTotal,
       unpaidFundingTotal: unpaidFundingTotal,
     };
@@ -524,20 +655,21 @@ export class RiskCalculator {
    * @param marginAccount   the user's MarginAccount.
    */
   public getMarginAccountState(
-    marginAccount: MarginAccount,
-    pnlExecutionPrice: number = undefined,
-    pnlAddTakerFees: boolean = false
+    marginAccount: MarginAccount
   ): types.MarginAccountState {
     let balance = convertNativeBNToDecimal(marginAccount.balance);
     let unrealizedPnl = this.calculateUnrealizedPnl(
       marginAccount,
-      types.ProgramAccountType.MarginAccount,
-      pnlExecutionPrice,
-      pnlAddTakerFees
+      types.ProgramAccountType.MarginAccount
     ) as number;
-    let unpaidFunding = this.calculateUnpaidFunding(marginAccount) as number;
+    let unpaidFunding = this.calculateUnpaidFunding(
+      marginAccount,
+      types.ProgramAccountType.MarginAccount
+    ) as number;
     let initialMargin = this.calculateTotalInitialMargin(
-      marginAccount
+      marginAccount,
+      types.ProgramAccountType.MarginAccount,
+      false
     ) as number;
     let initialMarginSkipConcession = this.calculateTotalInitialMargin(
       marginAccount,
@@ -545,7 +677,8 @@ export class RiskCalculator {
       true
     ) as number;
     let maintenanceMargin = this.calculateTotalMaintenanceMargin(
-      marginAccount
+      marginAccount,
+      types.ProgramAccountType.MarginAccount
     ) as number;
     let availableBalanceInitial: number =
       balance + unrealizedPnl + unpaidFunding - initialMargin;
@@ -564,5 +697,137 @@ export class RiskCalculator {
       availableBalanceMaintenance,
       availableBalanceWithdrawable,
     };
+  }
+
+  /**
+   * Find the maximum size a user is allowed to trade, given the position they're currently in
+   * @param marginAccount The user's CrossMarginAccount itself
+   * @param tradeAsset The asset being traded
+   * @param tradeSide The side (bid/ask) being traded
+   * @param tradePrice The price the user wishes to execute at
+   * @param bufferPercent An added buffer on top of the final result, so that quick price movements don't immediately make you hit leverage limits
+   * @param addTakerFees Whether to account for taker fees or not in the estimations
+   * @returns
+   */
+  public getMaxTradeSize(
+    marginAccount: CrossMarginAccount,
+    tradeAsset: constants.Asset,
+    tradeSide: types.Side,
+    tradePrice: number,
+    bufferPercent: number = 5,
+    addTakerFees: boolean = true
+  ): number {
+    let state = this.getCrossMarginAccountState(marginAccount);
+
+    let executionPrices = new Map();
+    executionPrices.set(tradeAsset, tradePrice);
+
+    let estimatedPnl = this.estimateRealizedPnl(
+      marginAccount,
+      types.ProgramAccountType.CrossMarginAccount,
+      { asset: tradeAsset, price: tradePrice, addTakerFees: addTakerFees }
+    );
+    let realizedBalanceInit =
+      state.balance +
+      estimatedPnl +
+      state.unpaidFundingTotal -
+      state.initialMarginTotal;
+
+    let realizedBalanceMaintInclOrders =
+      state.balance +
+      estimatedPnl +
+      state.unpaidFundingTotal -
+      state.maintenanceMarginIncludingOrdersTotal;
+
+    let fee = addTakerFees
+      ? (convertNativeBNToDecimal(Exchange.state.nativeD1TradeFeePercentage) /
+          100) *
+        tradePrice
+      : 0;
+
+    let position = convertNativeLotSizeToDecimal(
+      marginAccount.productLedgers[assets.assetToIndex(tradeAsset)].position
+        .size
+    );
+    let positionAbs = Math.abs(position);
+    let init = convertNativeBNToDecimal(
+      Exchange.pricing.marginParameters[assets.assetToIndex(tradeAsset)]
+        .futureMarginInitial,
+      constants.MARGIN_PRECISION
+    );
+    let maint = convertNativeBNToDecimal(
+      Exchange.pricing.marginParameters[assets.assetToIndex(tradeAsset)]
+        .futureMarginMaintenance,
+      constants.MARGIN_PRECISION
+    );
+    let markPrice = Exchange.getMarkPrice(tradeAsset);
+    let extraUPNLPerLot =
+      tradeSide == types.Side.BID
+        ? markPrice - tradePrice
+        : tradePrice - markPrice;
+
+    // (currentBalance + currentUPNL + unpaid funding - currentIM) + extraUPNL - extraIM - fees > 0
+    // state.availableBalanceInitial + extraUPNL - extraIM - fees > 0
+    // Where X is the number of extra lots:
+    // state.availableBalanceInitial + X*extraUPNLPerLot - X*init*markPrice - X*fee > 0
+    // X < state.availableBalanceInitial / (init*markPrice + fee - extraUPNLPerLot)
+    let openSize =
+      state.availableBalanceInitial /
+      (init * markPrice + fee - extraUPNLPerLot);
+
+    // (currentBalance + currentUPNL + unpaid funding - currentMMIO) + extraMMIO - fees > 0
+    // realizedBalanceMaintInclOrders + extraMMIO - fees > 0
+    // Where X is the number of extra lots:
+    // realizedBalanceMaintInclOrders + X*maint*markPrice - X*fee > 0
+    // X < realizedBalanceMaintInclOrders / (fee - maint*markPrice)
+    let closeSize = realizedBalanceMaintInclOrders / (fee - maint * markPrice);
+
+    // Equation falls apart if there is no positive X, means that you can close your position entirely
+    if (closeSize <= 0 || closeSize >= positionAbs) {
+      closeSize = positionAbs;
+    }
+
+    // (currentBalance + currentUPNL + unpaid funding - currentIM) + extraIMClose - extraIMOpen + extraUPNLOpen - fees > 0
+    // realizedBalanceInit + extraIMClose - extraIMOpen + extraUPNLOpen - fees > 0
+    // Where X is the number of extra lots opened on the other side:
+    // realizedBalanceInit + position*init*markPrice - X*init*markPrice + X*extraUPNLPerLot - X*fee > 0
+    // X < (realizedBalanceInit + position*init*markPrice) / (init*markPrice + fee - extraUPNLPerLot)
+
+    let closeOpenSize = closeSize;
+    if (closeSize == positionAbs) {
+      closeOpenSize =
+        positionAbs +
+        (realizedBalanceInit + positionAbs * init * markPrice) /
+          (init * markPrice + fee - extraUPNLPerLot);
+    }
+
+    if (
+      position == 0 ||
+      (position > 0 && tradeSide == types.Side.BID) ||
+      (position < 0 && tradeSide == types.Side.ASK)
+    ) {
+      // Strictly an opening order - use initial margin
+      return ((100 - bufferPercent) / 100) * openSize;
+    } else {
+      // Either only close (uses maint margin incl orders),
+      // or close full size + open more (uses initial margin)
+      // Return the one that gives the biggest size
+      return Math.max(((100 - bufferPercent) / 100) * closeOpenSize, closeSize);
+    }
+  }
+
+  public getLiquidationPrice(
+    asset: Asset,
+    signedPosition: number,
+    marginAccount: CrossMarginAccount
+  ) {
+    let state = this.getCrossMarginAccountState(marginAccount);
+    return calculateLiquidationPrice(
+      state.balance,
+      state.maintenanceMarginTotal,
+      state.unrealizedPnlTotal,
+      Exchange.getMarkPrice(asset),
+      signedPosition
+    );
   }
 }