+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Version changes are pinned to SDK releases.
 
 ## Unreleased
 
+## [1.6.0] 2023-07-25
+
+- Add getMaxTradeSize() and getLiquidationPrice() to risk calcs. ([#252](https://github.com/zetamarkets/sdk/pull/252))
+- Reformat pnl into estimateRealizedPnl() and calculateUnrealizedPnl(), using function overloads ([#252](https://github.com/zetamarkets/sdk/pull/252))
+- Add checkLiquidity() util function to find the best price in the orderbook for a given size. ([#252](https://github.com/zetamarkets/sdk/pull/252))
+
 ## [1.5.0] 2023-07-25
 
 - Clean up some risk calculations under the hood. ([#255](https://github.com/zetamarkets/sdk/pull/255))